+++ e2e-tests/src/main/java/gov/nasa/jpl/aerie/e2e/procedural/constraints/external_events/ExternalEventAttributeConstraint.java
@@ -1,21 +1,29 @@
-package gov.nasa.jpl.aerie.e2e.procedural.scheduling.procedures;
+package gov.nasa.jpl.aerie.e2e.procedural.constraints.external_events;
 
 import gov.nasa.ammos.aerie.procedural.constraints.Constraint;
 import gov.nasa.ammos.aerie.procedural.constraints.Violation;
 import gov.nasa.ammos.aerie.procedural.constraints.Violations;
 import gov.nasa.ammos.aerie.procedural.constraints.annotations.ConstraintProcedure;
+import gov.nasa.ammos.aerie.procedural.timeline.plan.EventQuery;
 import gov.nasa.ammos.aerie.procedural.timeline.plan.Plan;
 import gov.nasa.ammos.aerie.procedural.timeline.plan.SimulationResults;
 import org.jetbrains.annotations.NotNull;
 
 @ConstraintProcedure
-public record ExternalEventAbsenceConstraint() implements Constraint {
+public record ExternalEventAttributeConstraint(String codeValue) implements Constraint {
   @NotNull
   @Override
   public Violations run(@NotNull Plan plan, @NotNull SimulationResults simResults) {
-    var events = plan.events();
+    // filter events on whether attribute "code" equals codeValue
+    var events = plan.events(new EventQuery(null, "TestType", null))
+        .filter(true, e -> {
+          if (e.attributes.get("code").asString().isPresent()) {
+            return e.attributes.get("code").asString().get().equals(codeValue);
+          }
+          return false;
+        });
 
-    // if we have any events
+    // if we have any events with that attribute value
     if (!events.collect().isEmpty()) {
       // ...mark those events as violating
       return new Violations(
