+++ e2e-tests/src/main/java/gov/nasa/jpl/aerie/e2e/procedural/constraints/external_events/ExternalEventPresenceConstraint.java
@@ -1,4 +1,4 @@
-package gov.nasa.jpl.aerie.e2e.procedural.scheduling.procedures;
+package gov.nasa.jpl.aerie.e2e.procedural.constraints.external_events;
 
 import gov.nasa.ammos.aerie.procedural.constraints.Constraint;
 import gov.nasa.ammos.aerie.procedural.constraints.Violation;
