+++ e2e-tests/src/main/java/gov/nasa/jpl/aerie/e2e/procedural/goals/AnchorCascadeDeleteGoal.java
@@ -1,4 +1,4 @@
-package gov.nasa.jpl.aerie.e2e.procedural.scheduling.procedures;
+package gov.nasa.jpl.aerie.e2e.procedural.goals;
 
 import gov.nasa.ammos.aerie.procedural.scheduling.ActivityAutoDelete;
 import gov.nasa.ammos.aerie.procedural.scheduling.Goal;
@@ -10,32 +10,43 @@ import gov.nasa.ammos.aerie.procedural.timeline.plan.Plan;
 import gov.nasa.ammos.aerie.procedural.timeline.plan.SimulationResults;
 import gov.nasa.jpl.aerie.merlin.protocol.types.Duration;
 import gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue;
+import gov.nasa.jpl.aerie.types.ActivityDirectiveId;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.Map;
 
-/**
- * Creates one activity, and deletes it automatically on subsequent runs.
- */
 @SchedulingProcedure
-public record ActivityAutoDeletionGoal(boolean deleteAtBeginning) implements Goal {
+public record AnchorCascadeDeleteGoal() implements Goal {
+
   @NotNull
   @Override
   public ActivityAutoDelete shouldDeletePastCreations(
       @NotNull final Plan plan,
-      @Nullable final SimulationResults simResults)
-  {
-    if (deleteAtBeginning) return new ActivityAutoDelete.AtBeginning(DeletedAnchorStrategy.Error, false);
-    else return new ActivityAutoDelete.JustBefore(DeletedAnchorStrategy.Error);
+      @Nullable final SimulationResults simResults) {
+    // Delete activities created by previous runs of this goal, using Cascade to
+    // handle anchored activities
+    return new ActivityAutoDelete.AtBeginning(DeletedAnchorStrategy.Cascade, false);
   }
 
   @Override
   public void run(@NotNull final EditablePlan plan) {
-    plan.create(
+    final var ids = new ActivityDirectiveId[3];
+
+    ids[0] = plan.create(
         "BiteBanana",
-        new DirectiveStart.Absolute(Duration.MINUTE),
+        new DirectiveStart.Absolute(Duration.HOUR),
         Map.of("biteSize", SerializedValue.of(1))
+    );
+    ids[1] = plan.create(
+        "BiteBanana",
+        new DirectiveStart.Anchor(ids[0], Duration.HOUR, DirectiveStart.Anchor.AnchorPoint.End),
+        Map.of("biteSize", SerializedValue.of(2))
+    );
+    ids[2] = plan.create(
+        "BiteBanana",
+        new DirectiveStart.Anchor(ids[1], Duration.HOUR, DirectiveStart.Anchor.AnchorPoint.Start),
+        Map.of("biteSize", SerializedValue.of(3))
     );
 
     plan.commit();
