+++ e2e-tests/src/main/java/gov/nasa/jpl/aerie/e2e/procedural/goals/DeleteBiteBananasGoal.java
@@ -1,16 +1,10 @@
-package gov.nasa.jpl.aerie.e2e.procedural.scheduling.procedures;
+package gov.nasa.jpl.aerie.e2e.procedural.goals;
 
 import gov.nasa.ammos.aerie.procedural.scheduling.Goal;
 import gov.nasa.ammos.aerie.procedural.scheduling.annotations.SchedulingProcedure;
 import gov.nasa.ammos.aerie.procedural.scheduling.plan.DeletedAnchorStrategy;
 import gov.nasa.ammos.aerie.procedural.scheduling.plan.EditablePlan;
-import gov.nasa.ammos.aerie.procedural.timeline.payloads.activities.DirectiveStart;
-import gov.nasa.jpl.aerie.merlin.protocol.types.Duration;
-import gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue;
-import gov.nasa.jpl.aerie.types.ActivityDirectiveId;
 import org.jetbrains.annotations.NotNull;
-
-import java.util.Map;
 
 /**
  * Deletes all Bite Bananas with extreme prejudice. Used to test that updated
