+++ e2e-tests/src/main/java/gov/nasa/jpl/aerie/e2e/procedural/goals/external_events/ExternalEventsSourceQueryGoal.java
@@ -1,9 +1,8 @@
-package gov.nasa.jpl.aerie.e2e.procedural.scheduling.procedures;
+package gov.nasa.jpl.aerie.e2e.procedural.goals.external_events;
 
 import gov.nasa.ammos.aerie.procedural.scheduling.Goal;
 import gov.nasa.ammos.aerie.procedural.scheduling.annotations.SchedulingProcedure;
 import gov.nasa.ammos.aerie.procedural.scheduling.plan.EditablePlan;
-import gov.nasa.ammos.aerie.procedural.timeline.payloads.ExternalSource;
 import gov.nasa.ammos.aerie.procedural.timeline.payloads.activities.DirectiveStart;
 import gov.nasa.ammos.aerie.procedural.timeline.plan.EventQuery;
 import gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue;
@@ -13,14 +12,20 @@ import java.util.List;
 import java.util.Map;
 
 @SchedulingProcedure
-public record ExternalEventsEventAttributeQueryGoal() implements Goal {
+public record ExternalEventsSourceQueryGoal() implements Goal {
   @Override
   public void run(@NotNull final EditablePlan plan) {
-    // extract all events
-    for (final var e: plan.events()) {
-      // filter events that we schedule off of by their source's attributes
-      var version = e.attributes.get("projectUser").asString();
-      if (version.isPresent() && version.get().equals("UserA")) {
+
+    // extract events belonging to the second source
+    EventQuery eventQuery = new EventQuery(
+        null,
+        null,
+        List.of(new EventQuery.SourceQuery("NewTest.json", "TestGroup_2"))
+    );
+
+    for (final var e: plan.events(eventQuery)) {
+      // filter events that we schedule off of by key
+      if (e.key.contains("01")) {
         plan.create(
             "BiteBanana",
             // place the directive such that it is coincident with the event's start
