+++ e2e-tests/src/main/java/gov/nasa/jpl/aerie/e2e/procedural/package-info.java
@@ -1,5 +1,5 @@
 @WithMappers(BasicValueMappers.class)
-package gov.nasa.jpl.aerie.e2e.procedural.scheduling;
+package gov.nasa.jpl.aerie.e2e.procedural;
 
 import gov.nasa.jpl.aerie.contrib.serialization.rulesets.BasicValueMappers;
 import gov.nasa.ammos.aerie.procedural.scheduling.annotations.WithMappers;
