--- e2e-tests/src/main/java/gov/nasa/jpl/aerie/e2e/procedural/scheduling/procedures/ExternalEventsEventAttributeOptionalQueryGoal.java
@@ -1,33 +1,0 @@
-package gov.nasa.jpl.aerie.e2e.procedural.scheduling.procedures;
-
-import gov.nasa.ammos.aerie.procedural.scheduling.Goal;
-import gov.nasa.ammos.aerie.procedural.scheduling.annotations.SchedulingProcedure;
-import gov.nasa.ammos.aerie.procedural.scheduling.plan.EditablePlan;
-import gov.nasa.ammos.aerie.procedural.timeline.payloads.activities.DirectiveStart;
-import gov.nasa.jpl.aerie.merlin.protocol.types.SerializedValue;
-import org.jetbrains.annotations.NotNull;
-
-import java.util.Map;
-
-@SchedulingProcedure
-public record ExternalEventsEventAttributeOptionalQueryGoal() implements Goal {
-  @Override
-  public void run(@NotNull final EditablePlan plan) {
-    // extract all events
-    for (final var e: plan.events()) {
-      // filter events that we schedule off of by their attributes
-      var optionalValue = e.attributes.get("optional");
-      if (optionalValue != null) {
-        var optional = optionalValue.asString();
-        if (optional.isPresent() && optional.get().equals("present")) {
-          plan.create(
-              "BiteBanana",
-              // place the directive such that it is coincident with the event's start
-              new DirectiveStart.Absolute(e.getInterval().start),
-              Map.of("biteSize", SerializedValue.of(1)));
-        }
-      }
-    }
-    plan.commit();
-  }
-}
